requirements.txt
pandas
numpy
tensorflow
audioread >= 2.0.0
scipy >= 1.0.0
scikit-learn >= 0.14.0, != 0.19.0
joblib >= 0.12
decorator >= 3.0.0
six >= 1.3
resampy >= 0.2.2
numba >= 0.43.0
soundfile >= 0.9.0
librosa